pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "feedparser>=6.0.11",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-sqlalchemy>=3.1.1",
    "gevent>=25.5.1",
    "gunicorn>=23.0.0",
    "numpy<2.0",
    "openai>=1.98.0",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "sqlalchemy>=2.0.42",
    "ta>=0.11.0",
    "tensorflow>=2.14.0",
    "trafilatura>=2.0.0",
    "xgboost>=3.0.3",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
]
